requirements.txt
# Airflow (installed separately via Docker, listed here for reference)
apache-airflow==2.8.1

# Azure
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Snowflake
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1

# dbt
dbt-core==1.7.0
dbt-snowflake==1.7.0

# Data processing
pandas==2.2.0
requests==2.31.0

# Kafka
confluent-kafka==2.3.0

# Utilities
python-dotenv==1.0.0